Fix test fixture requiring standardize_response outside of request
One of the tests in test_routes.py utilizes a fixture in conftest.py
that relies on standardize_response to work outside of a Flask request
context. I'm pretty sure the test shouldn't rely on it in that way, but
this was the easiest path forward to a clean CI/CD build.